Más legible

+++ Funciones-Test.h
@@ -2,7 +2,7 @@ vector<Libro> genLibros(int Nlibros){
     vector<Libro> resultado;
     string nom;
 	int id;
-	for(size_t i = 0; i<Nlibros; ++i){
+	for(int i = 0; i<Nlibros; ++i){
         nom= "test "+to_string(i+1);
 		id= rand()%(Nlibros*2)+1;
 		Libro aux(id,nom.c_str());
